+++ packages/types/src/config.ts
@@ -121,8 +121,10 @@ export type ClientConfig = {
     FeatureFlagBoardsProduct: string;
     FeatureFlagCallsEnabled: string;
     FeatureFlagGraphQL: string;
+    FeatureFlagIkCallDialing: string;
     GfycatAPIKey: string;
     GfycatAPISecret: string;
+    GiphySdkKey: string;
     GoogleDeveloperKey: string;
     GuestAccountsEnforceMultifactorAuthentication: string;
     HasImageProxy: string;
@@ -326,6 +328,7 @@ export type ServiceSettings = {
     EnableGifPicker: boolean;
     GfycatAPIKey: string;
     GfycatAPISecret: string;
+    GiphySdkKey: string;
     PostEditTimeLimit: number;
     TimeBetweenUserTypingUpdatesMilliseconds: number;
     EnablePostSearch: boolean;