+++ reducers/views/calls.ts
@@ -5,6 +5,9 @@ import {combineReducers} from 'redux';
 import {UserProfile} from 'mattermost-redux/types/users';
 
 import {ActionTypes} from 'utils/constants';
+import {Post} from '@mattermost/types/posts';
+import {ServerChannel} from '@mattermost/types/channels';
+import {GenericAction} from 'mattermost-redux/types/actions';
 
 export type UserState = {
     voice: boolean;
@@ -425,6 +428,30 @@ const callStartAt = (state: {[channelID: string]: number} = {}, action: {type: s
     }
 };
 
+const callParameters = (
+    state: {users: UserProfile[]; caller: UserProfile; channel: ServerChannel; msg: Post} = {users: [], caller: {}, channel: {}, msg: {}},
+    action: GenericAction) => {
+    switch (action.type) {
+    case ActionTypes.CALL_USERS_IN_CONF:
+        return {
+            ...state, users: action.data,
+        };
+    case ActionTypes.CALL_CALLING_USER:
+        return {
+            ...state, caller: action.data[0],
+        };
+    case ActionTypes.CALL_CONF_CHANNEL:
+        return {
+            ...state, channel: action.data,
+        };
+    case ActionTypes.CALL_RECEIVED:
+        return {
+            ...state, msg: action.data.msg,
+        };
+    default:
+        return state;
+    }
+};
 const voiceChannelScreenSharingID = (state: {[channelID: string]: string} = {}, action: {type: string; data: {channelID: string; userID?: string}}) => {
     switch (action.type) {
     case ActionTypes.VOICE_CHANNEL_UNINIT:
@@ -496,4 +523,5 @@ export default combineReducers({
     expandedView,
     switchCallModal,
     screenSourceModal,
+    callParameters,
 });