+++ selectors/calls.ts
@@ -1,12 +1,36 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
-import {getCurrentChannelId} from 'mattermost-redux/selectors/entities/common';
+
+import {getCurrentChannelId, getCurrentUserId} from 'mattermost-redux/selectors/entities/common';
 import {GlobalState} from 'types/store';
 
 export const connectedChannelID = (state: GlobalState) => state.views.calls.connectedChannelID;
 export const connectedCallID = (state: GlobalState) => state.views.calls.connectedCallID;
 export const connectedCallUrl = (state: GlobalState) => state.views.calls.connectedCallUrl;
 export const voiceConnectedChannels = (state: GlobalState) => state.views.calls.voiceConnectedChannels;
+
+export const usersInCall = (state: GlobalState) => state.views.calls.callParameters.users;
+export const userCalling = (state: GlobalState) => state.views.calls.callParameters.caller;
+export const channelInCall = (state: GlobalState) => state.views.calls.callParameters.channel;
+export const callConferenceId = (state: GlobalState) => state.views.calls.callParameters.msg.props.conference_id;
+export const callMessage = (state: GlobalState) => state.views.calls.callParameters.msg;
+
+export const userStatus = (state: GlobalState) => state.entities.users.statuses;
+
+export const callParameters = (state: GlobalState) => {
+    return {
+        users: usersInCall(state).filter((usr) => usr.id !== getCurrentUserId(state)),
+        caller: userCalling(state),
+        channel: channelInCall(state),
+        msg: callMessage(state),
+    };
+};
+
+export const callUserStatus = (state: GlobalState) => {
+    return {
+        status: userStatus(state),
+    };
+};
 
 export const voiceConnectedUsers = (state: GlobalState) => {
     const currentChannelID = getCurrentChannelId(state);