+++ utils/utils.tsx
@@ -58,6 +58,7 @@ import crackle from 'sounds/crackle.mp3';
 import down from 'sounds/down.mp3';
 import hello from 'sounds/hello.mp3';
 import ripple from 'sounds/ripple.mp3';
+import ring from 'sounds/ring.mp3';
 import upstairs from 'sounds/upstairs.mp3';
 import {t} from 'utils/i18n';
 import store from 'stores/redux_store.jsx';
@@ -229,6 +230,7 @@ export const notificationSounds = new Map([
     ['Hello', hello],
     ['Ripple', ripple],
     ['Upstairs', upstairs],
+    ['Ring', ring],
 ]);
 
 let canDing = true;
@@ -1942,3 +1944,25 @@ export function getBlankAddressWithCountry(country?: string): Address {
         state: '',
     };
 }
+
+export const lazyWithRetries: typeof React.lazy = (importer) => {
+    const retryImport = async () => {
+        try {
+            return await importer();
+        } catch (error) {
+            for (let i = 0; i < 5; i++) {
+                // eslint-disable-next-line no-await-in-loop
+                await new Promise((resolve) => setTimeout(resolve, 2000));
+                try {
+                    // eslint-disable-next-line no-await-in-loop
+                    return await importer();
+                } catch (e) {
+                    // eslint-disable-next-line no-console
+                    console.log('retrying import');
+                }
+            }
+            throw error;
+        }
+    };
+    return React.lazy(retryImport);
+};