+++ webpack.config.js
@@ -274,7 +274,12 @@ var config = {
             'mattermost-redux': 'packages/mattermost-redux/src',
             reselect: 'packages/reselect/src',
             marked: '@infomaniak/marked',
+            '@mattermost/client': '@infomaniak/mattermost-client',
+            '@mattermost/types': '@infomaniak/mattermost-types',
             '@mui/styled-engine': '@mui/styled-engine-sc',
+
+            // This alias restricts single version of styled components acros all packages
+            'styled-components': path.resolve(__dirname, '.', 'node_modules', 'styled-components'),
         },
         extensions: ['.ts', '.tsx', '.js', '.jsx'],
         fallback: {